[Web] Refactor: :rainbow: rename dataType file and define enum socketErrorMessage #61
dataType.ts 파일의 이름을 수정하고, socketErrorMessage를 enum으로 선언
