+++ server/src/@types/socketErrorMessage.ts
@@ -1,0 +1,9 @@
+enum SocketErrorMessage {
+  ROOM_CODE = 'Room code is not valid',
+  NICKNAME = 'Nickname is not valid',
+  LANGUAGE = 'Language is not valid',
+  CHAT = 'Chat message is not valid',
+  SERVER = 'Internal server error',
+}
+
+export default SocketErrorMessage;
