+++ server/src/@types/statusCode.ts
@@ -1,6 +1,8 @@
 enum StatusCode {
   OK = 200,
   CLIENT_ERROR = 400,
+  SERVER_ERROR = 500,
+  NOT_ACCEPTABLE = 406,
 }
 
 export default StatusCode;
